backend/requirements.txt
flask
flask-sqlalchemy
flask-jwt-extended
flask-cors
werkzeug
flask-bcrypt
pandas
openpyxl
